minor menu error
Issue occurs after removing a movie from the shopping cart. The viewCart
function would call a delete function which would again call the
viewCart function. If I tried to quit right afterwards, this would
result in quitting the viewCart function but only to immediately return
to it. Instead of having removeCart call viewCart again I put a label at
the top of viewCart and called it after calling removeCart.